refactor: reuse AlliumMenu instead of respawning
The initialization of AlliumMenu does some heavy work, such as loading
the theme and locale. By keeping a persistent AlliumMenu instance in
AlliumD, we reduce the launch delay at the cost of some memory overhead.

The additional memory used is fine because when AlliumMenu is launched
anyway, RetroArch will still be active and the maximum memory used is
identical.